QrCode-tri-recherche-email
update(not final version)

+++ src/main/java/tn/esprit/biol/service/EquipmentService.java
@@ -1,18 +1,28 @@
 package tn.esprit.biol.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.PathVariable;
 import tn.esprit.biol.dao.EquipmentDao;
 import tn.esprit.biol.entity.Equipment;
 
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+
 
 @Service
 public class EquipmentService implements IequipmentService{
     @Autowired
+    EmailServiceSender es;
+    @Autowired
     private EquipmentDao EqRepo;
+
     @Override
     public List<Equipment> getAllEquipments() {
         List<Equipment> equipment = new ArrayList<>();
@@ -20,9 +30,26 @@ public class EquipmentService implements IequipmentService{
         return equipment;
     }
 
-    @Override
+   /* @Override
     public Equipment AddEquipment(Equipment e) {
         return EqRepo.save(e);
+    }*/
+
+    @Override
+    public ResponseEntity<?> AddEq(Equipment equipment) {
+
+        if (equipment.getType().isEmpty() || equipment.getType().equals(0)) {
+            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("Choose a correct type above 0");
+        } else if (equipment.getExpiration_Date() == null || (equipment.getExpiration_Date().isEqual(LocalDate.now())) || equipment.getExpiration_Date().isBefore(LocalDate.now())) {
+            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("DATE NOT ACCEPTABLE");
+        } else if (equipment.getSterilization_Date() == null || (equipment.getSterilization_Date().isEqual(LocalDate.now())) || equipment.getSterilization_Date().isBefore(LocalDate.now())) {
+            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("DATE NOT ACCEPTABLE");
+        } else if (equipment.getQuantity() == null || equipment.getQuantity() < 1) {
+            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("Quantity can't be null");
+        } else {
+            EqRepo.save(equipment);
+            return ResponseEntity.status(HttpStatus.CREATED).body(equipment);
+        }
     }
 
     @Override
@@ -31,10 +58,19 @@ public class EquipmentService implements IequipmentService{
     }
 
     @Override
-    public void deleteEquipment(Integer Id_eq) {
-        EqRepo.deleteById(Id_eq);
+    public ResponseEntity<?> deleteEq(Integer Id_eq ) {
 
-    }
+        Optional<Equipment> equipment = EqRepo.findById(Id_eq);
+        if(!equipment.isPresent()) {
+            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("there is no such Equipment with this id: "+Id_eq);
+        }
+        else
+        {
+
+        EqRepo.deleteById(Id_eq);
+        return ResponseEntity.status(HttpStatus.OK).body("Equipment deleted");
+
+    }}
 
     @Override
     public Equipment updateEquipment(Equipment e,Integer idEq) {
@@ -45,5 +81,40 @@ public class EquipmentService implements IequipmentService{
         eq.setSterilization_Date(e.getSterilization_Date());
         return EqRepo.save(eq);
     }
+
+    @Override
+    public void sendSterilizationRequest(String to,String f,String l) {
+        // Envoi de l'email de demande de congé
+        es.sendEmail(to,f,l);
+    }
+    @Override
+
+    public void sendLowStorageRequest(String to,String sub,String bod) {
+        // Envoi de l'email de demande de congé
+        es.sendEmail(to,sub,bod);
+    }
+    @Override
+
+    public List<Equipment> getAllEquipmentswithASCSorting(String field){
+        return EqRepo.findAll(Sort.by(Sort.Direction.ASC,field));
+    }
+    @Override
+
+    public List<Equipment> getAllEquipmentswithDESCSorting(String field){
+        return EqRepo.findAll(Sort.by(Sort.Direction.DESC,field));
+    }
+    @Override
+
+    public Page<Equipment> getAllEquipmentswithPagination(Integer offset,Integer pageSize){
+        Page<Equipment> equip= EqRepo.findAll(PageRequest.of(offset, pageSize));
+        return equip;
+    }
+
+    public Equipment getEquipment(Integer id) {
+        return EqRepo.findById(id).get();
+
+    }
+
+
 
 }
